[feature/11] 메인페이지 콘텐츠/추천 탭 생성 및 기타 css 수정
