+++ src/views/MBTIny/style/main.css
@@ -1,108 +1,34 @@
-/* Main container styling */
-.recommd-main-container {
-    width: 100%; 
-    overflow-x: hidden;
-}
-
-.recommd-content-container {
+/* 탭 메뉴 스타일 */
+.tab-menu {
     display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    background-color: #FFE7EB;
-    padding: 20px;
+    justify-content: center; /* 왼쪽 정렬 */
+    background-color: #FFE7EB; /* 배경을 부드러운 색으로 */
+    padding: 10px 20px;
 }
 
-.recommended-books {
-    padding: 60px;
-}
-
-.recommended-books h3{
-    margin-bottom: 5px;
-}
-
-.recommended-books p{
-    padding-left: 20px; /* 원하는 만큼의 들여쓰기 설정 */
-    font-size: 12px; /* 적절한 크기로 폰트 조절 */
-    color:rgb(242, 58, 128);
-}
-
-.book-container {
-    gap: 15px;
-    display: flex;
-    flex-direction: row;
-    overflow-x: auto;
-    white-space: nowrap;
-    scrollbar-width: none; /* Firefox */
-    -ms-overflow-style: none; /* IE and Edge */
-    scroll-behavior: smooth;
-    padding: 20px 20px 0;
-    width: 1550px;
-}
-
-.book-container::-webkit-scrollbar {
-    display: none; /* Chrome, Safari, Opera */
-}
-
-.book-item {
-    min-width: 180px;
-    max-width: 180px;
-    text-align: center;
-    flex-shrink: 0;
-    
-}
-
-.book-cover {
-    width: 100%;
-    height: auto;
-    border-radius: 10px;
-    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.3);
-    transition: transform 0.2s ease;
-}
-
-.book-cover:hover {
-    transform: scale(1.05);
-}
-
-.book-item p {
-    margin-top: 5px;
-    font-size: 0.8rem;
-    color: #333;
-    font-weight: 100;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-/* Scroll button styling */
-.scroll-button {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    background-color: rgba(0, 0, 0, 0.5);
-    color: white;
-    border: none;
-    border-radius: 50%;
-    width: 40px;
-    height: 40px;
-    font-size: 1.5rem;
+.tab-menu button {
+    padding: 10px 20px;
+    margin-right: 10px;
+    border-radius: 20px;
+    background-color: #FFDDE2; /* 부드러운 핑크색 배경 */
+    color: #666;
+    font-weight: 500;
     cursor: pointer;
-    z-index: 1;
+    font-size: 16px;
+    border: none;
+    transition: background-color 0.3s ease, color 0.3s ease;
+    font-family: 'UhBeeSe_hyun';
+    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1); /* 가벼운 그림자 */
 }
 
-.scroll-button.left {
-    left: 0;
+.tab-menu button.active-tab {
+    color: #fff; /* 선택된 탭 글씨 색을 흰색으로 */
+    background-color: #ff85a1; /* 선택된 탭 배경을 좀 더 진한 색으로 */
+    font-weight: bold;
+    box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.2); /* 선택된 탭에 더 깊은 그림자 */
 }
 
-.scroll-button.right {
-    right: 0;
-}
-
-.scroll-button:hover {
-    background-color: rgba(0, 0, 0, 0.7);
-}
-
-.recommd-container {
-    background-color: #FFE7EB;
-    height: var(—vh, 100vh);
-
+.tab-menu button:hover {
+    background-color: #FFC2CB; /* 호버 시 색상 변경 */
+    color: #fff;
 }
